# avoid sscanf to parse into double or identifier in lp reader